+++ libsimple_flutter/CHANGELOG.md
@@ -1,3 +1,17 @@
+## 0.0.5
+* change web sqlite run in worker in order to save db with OPFS
+
+## 0.0.4
+* update wasm version
+* add android support
+
+## 0.0.3
+* update version to sqlite(3.43.1) and libsimple(0.3.0-20230912)
+
+## 0.0.2 
+
+* add iOS support
+
 ## 0.0.1
 
-* TODO: Describe initial release.
+* a base init relase.